Replace go string equality with a scalar StringMatcher
